QABot/requirements.txt
langchain
langchain-community
streamlit
numpy
pandas
requests
langchain-google-genai
python-dotenv
unstructured 
libmagic 
python-magic 
Mysql-connector-python
pymysql
chromadb
langchain-experimental
faiss-cpu
python-magic-bin